Split ScopeStackCollector part into abstract class
